move shootingCreatureTest
