+++ src/test/java/pl/sdk/gui/CreatureTest.java
@@ -27,16 +27,4 @@ public class CreatureTest {
         //then
         assertEquals(1000, professor.getCurrentHp());
     }
-
-    @Test
-    void shouldNotCounterAttackWhenAttackerIsShooter(){
-        //given
-        Creature attacker = new ShootingCreature(100,1, 1, "Student",1,100);
-        Creature defender = new Creature(100,10, 1, "Student",1,100);
-        //when
-        attacker.attack(defender);
-        //then
-        assertEquals(100,attacker.getCurrentHp());
-        assertEquals(100,attacker.getAmount());
-    }
 }
